# src/main/resources/docs/api/Swagger.yml
openapi: 3.0.3
info:
  title: Sport Life API
  description: API for Sport Life training application
  version: 1.0.0

servers:
  - url: http://localhost:49182

components:

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Response:
      error:
        ErrorResponse:
          type: object
          properties:
            error:
              type: integer
              example: 401
            message:
              type: string
              example: Unauthorized
      calendar:
        InfoCalendarResponse:
          type: Object
          properties:
            monday:
              type: array
            tuesday:
              type: array
            wednesday:
              type: array
            thursday:
              type: array
            friday:
              type: array
            saturday:
              type: array
            sunday:
              type: array
        DeleteAndCreateCalendarResponse:
          type: Object
          properties:
            response:
              type: String
      employee:
        AuthAndCreateEmployeeResponse:
          type: Object
          properties:
            token:
              type: String
        UpdateAndUpdateActivityAndUpdateExpertsEmployeeResponse:
          type: Object
          properties:
            update:
              type: String
              example: Update Completed
        InfoEmployeeResponse:
          type: Object
          properties:
            login:
              type: String
            avatar:
              type: String
            experts:
              type: String
            activity:
              type: integer
        EmployeeTop:
          type: Object
          properties:
            login:
              type: String
            activity:
              type: integer
              format: int64
            avatar:
              type: String
        InfoTopEmployeeResponse:
          type: Object
          properties:
            top:
              type: Object
              items:
                $ref: '#/component/schemas/Response/EmployeeTop'
      exercise:
        ExerciseObject:
          type: Object
          properties:
            name:
              type: String
            video:
              type: String
            description:
              type: String
            photo:
              type: String
            males:
              type: array
            items:
              type: array
        InfoExerciseAndInfoFavouritesResponse:
          type: Object
          properties:
            exercises:
              type: array
              items:
                $ref: '#/component/schemas/Response/ExerciseObject'
      favourites:
        CreateAndDeleteFavouritesResponse:
          type: Object
          properties:
            response:
              type: String
    Request:
      calendar:
        DeleteAndCreateCalendarRequest:
          type: Object
          properties:
            name:
              type: String
            time:
              type: String
              format: time
      employee:
        AuthEmployeeRequest:
          type: object
          properties:
            login:
              type: string
            password:
              type: string
        CreateEmployeeRequest:
          type: Object
          properties:
            login:
              type: String
            password:
              type: String
            avatar:
              type: String
        UpdateExpertsEmployeeRequest:
          type: Object
          properties:
            experts:
              type: String
            login:
              type: String
        UpdateEmployeeRequest:
          type: Object
          properties:
            login:
              type: String
            avatar:
              type: String
      exercise:
        InfoExerciseRequest:
          type: Object
          properties:
            items:
              type: array
            males:
              type: array
      favourites:
        CreateAndDeleteFavouritesRequest:
          type: Object
          properties:
            name:
              type: String
            login:
              type: String
paths:
  /api/Calendar/info:
    get:
      summary: Get reminders
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List reminders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/calendar/InfoCalendarResponse'
        "404":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Calendar/create:
    post:
      summary: Create reminder
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/calendar/DeleteAndCreateCalendarRequest'
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/calendar/DeleteAndCreateCalendarResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Calendar/delete:
    delete:
      summary: Delete reminder
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/calendar/DeleteAndCreateCalendarRequest'
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/calendar/DeleteAndCreateCalendarResponse'
        "404":
          description: Not_Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'

  /api/Employee/auth:
    post:
      summary: Login employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/employee/AuthEmployeeRequest'
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/AuthAndCreateEmployeeResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "404":
          description: Not_Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/create:
    post:
      summary: Register employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/employee/CreateEmployeeRequest'
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/AuthAndCreateEmployeeResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "404":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/experts:
    patch:
      summary: Set user experience level
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/employee/UpdateExpertsEmployeeRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/UpdateAndUpdateActivityAndUpdateExpertsEmployeeResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/update:
    post:
      summary: Update profile
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/employee/UpdateEmployeeRequest'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/UpdateAndUpdateActivityAndUpdateExpertsEmployeeResponse'
        "404":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/info:
    get:
      summary: Get employee profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Employee infoFavourites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/InfoEmployeeResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "404":
          description: Not_Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/top:
    get:
      summary: Get rating leaderboard
      responses:
        "200":
          description: Top users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/InfoTopEmployeeResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Employee/activity:
    patch:
      summary: Increase employee activity score
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/employee/UpdateAndUpdateActivityAndUpdateExpertsEmployeeResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'

  /api/Exercise/info:
    post:
      summary: Search exercises
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/exercise/InfoExerciseRequest'
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/exercise/InfoExerciseAndInfoFavouritesResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'

  /api/Favourites/create:
    post:
      summary: Add exercise to favourites
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/favourites/CreateAndDeleteFavouritesRequest'
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/favourites/CreateAndDeleteFavouritesResponse'
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Favorites/delete:
    delete:
      summary: Remove exercise from favourites
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request/favourites/CreateAndDeleteFavouritesRequest'
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/favourites/CreateAndDeleteFavouritesResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
  /api/Favourites/info:
    get:
      summary: Get user exercise history
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Exercise history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/exercise/InfoExerciseAndInfoFavouritesResponse'
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response/error/ErrorResponse'